Use __init__ instead of hooks like other plugins
